requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.114.1
filelock==3.16.0
fsspec==2024.9.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.8
jiter==0.5.0
numpy==2.1.1
openai==1.45.0
packaging==24.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.38.5
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
